adds TipsRequester interface, use DedicatedScheduledExecutorService in TipsRequesterImpl
